feat(memory, var): add arena allocation, inference toggles, and mgrind leak tracing
- Add `tarena` memory allocator (`tensor_arena.c`/`.h`) for scoped allocations and resets.
- Update `tensor_memory` to support arena-backed allocations and conditional `mgrind` debugging.
- Add `mgrind` tool with `libbacktrace` support for allocation tracking and stack trace dumping.
- Add global and thread-local inference mode controls (`var_inference.c`).
- Improve safety across tensor ops (bounds checking in slices/gathers, stack-based odometer buffers, contiguous flags).
- Clean up autograd lifecycle in `var_backward` and `var_destroy` by recursively freeing gradients and tracking visited nodes.
- Update transformer training and inference pipeline test (`train_first_primer_and_infer.c`) to use arena resets.

+++ src/ops/tensor_random.c
@@ -38,7 +38,7 @@ void tensor_random(Tensor tensor) {
   f32 value;
   if (tensor->is_contiguous) {
     for (u32 i = 0; i < tensor->data->nelements; ++i) {
-      random_ratio = urand() / (f32)UINT32_MAX;
+      random_ratio = urand() / (f32)UINT64_MAX;
       tensor->data->data[i] =
           FLT_MIN + (FLT_MAX * random_ratio) - (FLT_MIN * random_ratio);
     }
